Move ChatModule into its own folder and move ChatChannel enum
