+++ TGM/src/main/java/network/warzone/tgm/command/CycleCommands.java
@@ -12,7 +12,8 @@ import network.warzone.tgm.map.MapContainer;
 import network.warzone.tgm.map.MapInfo;
 import network.warzone.tgm.match.MatchManager;
 import network.warzone.tgm.match.MatchStatus;
-import network.warzone.tgm.modules.ChatModule;
+import network.warzone.tgm.modules.chat.ChatChannel;
+import network.warzone.tgm.modules.chat.ChatModule;
 import network.warzone.tgm.modules.countdown.Countdown;
 import network.warzone.tgm.modules.countdown.CycleCountdown;
 import network.warzone.tgm.modules.countdown.StartCountdown;
@@ -554,10 +555,10 @@ public class CycleCommands {
         }
 
         String channelName = cmd.getString(0).toUpperCase();
-        ChatModule.Channel channel = ChatModule.Channel.byName(channelName);
+        ChatChannel channel = ChatChannel.byName(channelName);
         if (channel == null) {
           player.sendMessage(ColorConverter.filterString("&cInvalid channel: " + channelName));
-          player.sendMessage(ColorConverter.filterString("&cChannels: ( " + StringUtils.join(Arrays.stream(ChatModule.Channel.values()).filter(ch -> ch.hasPermission(player)).collect(Collectors.toList()), " | ")) + " )");
+          player.sendMessage(ColorConverter.filterString("&cChannels: ( " + StringUtils.join(Arrays.stream(ChatChannel.values()).filter(ch -> ch.hasPermission(player)).collect(Collectors.toList()), " | ")) + " )");
           return;
         }
