+++ TGM/src/main/java/network/warzone/tgm/match/MatchManifest.java
@@ -1,9 +1,9 @@
 package network.warzone.tgm.match;
 
 import com.google.gson.JsonObject;
-import network.warzone.tgm.map.MapInfo;
 import network.warzone.tgm.modules.*;
 import network.warzone.tgm.modules.border.WorldBorderModule;
+import network.warzone.tgm.modules.chat.ChatModule;
 import network.warzone.tgm.modules.countdown.CycleCountdown;
 import network.warzone.tgm.modules.countdown.StartCountdown;
 import network.warzone.tgm.modules.damage.DamageControlModule;
